requirements.txt
argon2-cffi==20.1.0
certifi==2019.11.28
cffi==1.14.2
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.8
oauthlib==3.1.0
PyJWT==1.7.1
pylint==2.5.3
pymongo==3.11.0
python-dotenv==0.14.0
requests==2.22.0
Werkzeug==1.0.0
wrapt==1.12.1

